rating/requirements.txt
torch
torchvision
pillow
redis
pymongo
requests